requirements.txt
fastapi
uvicorn
redis
pymongo
mysql-connector-python
pymysql